+++ emase/__init__.py
@@ -1,9 +1,10 @@
 # -*- coding: utf-8 -*-
+
 from .Sparse3DMatrix import Sparse3DMatrix
 from .AlignmentPropertyMatrix import AlignmentPropertyMatrix
 from .AlignmentMatrixFactory import AlignmentMatrixFactory
 from .EMfactory import EMfactory
 
-__author__ = 'Kwangbom "KB" Choi'
+__author__ = 'Kwangbom "KB" Choi, Ph. D., The Jackson Laboratory'
 __email__ = 'kb.choi@jax.org'
-__version__ = '0.9.4'
+__version__ = '0.10.11'